move TxConsequences to separate file
