+++ src/ripple/app/tx/impl/TxConsequences.cpp
@@ -1,0 +1,53 @@
+
+#include <ripple/app/tx/impl/TxConsequences.h>
+#include <ripple/app/main/Application.h>
+#include <ripple/basics/XRPAmount.h>
+#include <ripple/beast/utility/Journal.h>
+#include <ripple/core/Config.h>
+#include <ripple/ledger/ApplyViewImpl.h>
+#include <ripple/protocol/STTx.h>
+#include <optional>
+#include <utility>
+
+namespace ripple {
+
+TxConsequences::TxConsequences(NotTEC pfresult)
+    : isBlocker_(false)
+    , fee_(beast::zero)
+    , potentialSpend_(beast::zero)
+    , seqProx_(SeqProxy::sequence(0))
+    , sequencesConsumed_(0)
+{
+    assert(!isTesSuccess(pfresult));
+}
+
+TxConsequences::TxConsequences(STTx const& tx)
+    : isBlocker_(false)
+    , fee_(
+          tx[sfFee].native() && !tx[sfFee].negative() ? tx[sfFee].xrp()
+                                                      : beast::zero)
+    , potentialSpend_(beast::zero)
+    , seqProx_(tx.getSeqProxy())
+    , sequencesConsumed_(tx.getSeqProxy().isSeq() ? 1 : 0)
+{
+}
+
+TxConsequences::TxConsequences(STTx const& tx, Category category)
+    : TxConsequences(tx)
+{
+    isBlocker_ = (category == blocker);
+}
+
+TxConsequences::TxConsequences(STTx const& tx, XRPAmount potentialSpend)
+    : TxConsequences(tx)
+{
+    potentialSpend_ = potentialSpend;
+}
+
+TxConsequences::TxConsequences(STTx const& tx, std::uint32_t sequencesConsumed)
+    : TxConsequences(tx)
+{
+    sequencesConsumed_ = sequencesConsumed;
+}
+
+}
