+++ src/ripple/app/tx/impl/applySteps.cpp
@@ -317,45 +317,6 @@ invoke_calculateBaseFee(ReadView const& view, STTx const& tx)
     }
 }
 
-TxConsequences::TxConsequences(NotTEC pfresult)
-    : isBlocker_(false)
-    , fee_(beast::zero)
-    , potentialSpend_(beast::zero)
-    , seqProx_(SeqProxy::sequence(0))
-    , sequencesConsumed_(0)
-{
-    assert(!isTesSuccess(pfresult));
-}
-
-TxConsequences::TxConsequences(STTx const& tx)
-    : isBlocker_(false)
-    , fee_(
-          tx[sfFee].native() && !tx[sfFee].negative() ? tx[sfFee].xrp()
-                                                      : beast::zero)
-    , potentialSpend_(beast::zero)
-    , seqProx_(tx.getSeqProxy())
-    , sequencesConsumed_(tx.getSeqProxy().isSeq() ? 1 : 0)
-{
-}
-
-TxConsequences::TxConsequences(STTx const& tx, Category category)
-    : TxConsequences(tx)
-{
-    isBlocker_ = (category == blocker);
-}
-
-TxConsequences::TxConsequences(STTx const& tx, XRPAmount potentialSpend)
-    : TxConsequences(tx)
-{
-    potentialSpend_ = potentialSpend;
-}
-
-TxConsequences::TxConsequences(STTx const& tx, std::uint32_t sequencesConsumed)
-    : TxConsequences(tx)
-{
-    sequencesConsumed_ = sequencesConsumed;
-}
-
 static std::pair<TER, bool>
 invoke_apply(ApplyContext& ctx)
 {
